modeling/pyproject.toml
[project]
name = "market-benchmarking-modeling"
version = "0.1.0"
description = "Dataset inventory and data contract tooling for the explainable market benchmarking project."
requires-python = ">=3.13"
dependencies = [
  "pandas>=2.2",
  "pyarrow>=15.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=data_foundation --cov-report=term-missing --cov-fail-under=85"
